+++ lib/Job/PID.php
@@ -1,4 +1,9 @@
 <?php
+
+namespace Dynamo\Resque\Job;
+
+use Dynamo\Resque\Resque;
+
 /**
  * PID tracker for the forked worker job.
  *
@@ -6,7 +11,7 @@
  * @author		Chris Boulton <chris@bigcommerce.com>
  * @license		http://www.opensource.org/licenses/mit-license.php
  */
-class Resque_Job_PID
+class PID
 {
 	/**
 	 * Create a new PID tracker item for the supplied job ID.