+++ lib/Job/Status.php
@@ -1,4 +1,8 @@
 <?php
+
+namespace Dynamo\Resque\Job;
+
+use Dynamo\Resque\Resque;
 /**
  * Status tracker/information for a job.
  *
@@ -6,7 +10,7 @@
  * @author		Chris Boulton <chris@bigcommerce.com>
  * @license		http://www.opensource.org/licenses/mit-license.php
  */
-class Resque_Job_Status
+class Status
 {
 	const STATUS_WAITING = 1;
 	const STATUS_RUNNING = 2;