+++ test/Resque/Tests/LogTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Resque_Log tests.
+ * ConsoleLogger tests.
  *
  * @package		Resque/Tests
  * @author		Chris Boulton <chris@bigcommerce.com>
@@ -10,7 +10,7 @@ class Resque_Tests_LogTest extends Resque_Tests_TestCase
 {
 	public function testLogInterpolate()
 	{
-		$logger   = new Resque_Log();
+		$logger   = new ConsoleLogger();
 		$actual   = $logger->interpolate('string {replace}', array('replace' => 'value'));
 		$expected = 'string value';
 
@@ -19,7 +19,7 @@ class Resque_Tests_LogTest extends Resque_Tests_TestCase
 
 	public function testLogInterpolateMutiple()
 	{
-		$logger   = new Resque_Log();
+		$logger   = new ConsoleLogger();
 		$actual   = $logger->interpolate(
 			'string {replace1} {replace2}',
 			array('replace1' => 'value1', 'replace2' => 'value2')