+++ test/Resque/Tests/WorkerTest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Resque_Worker tests.
+ * Worker tests.
  *
  * @package		Resque/Tests
  * @author		Chris Boulton <chris@bigcommerce.com>
@@ -10,8 +10,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 {
 	public function testWorkerRegistersInList()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
 		// Make sure the worker is in the list
@@ -23,46 +23,46 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 		$num = 3;
 		// Register a few workers
 		for($i = 0; $i < $num; ++$i) {
-			$worker = new Resque_Worker('queue_' . $i);
-			$worker->setLogger(new Resque_Log());
+			$worker = new Worker('queue_' . $i);
+			$worker->setLogger(new ConsoleLogger());
 			$worker->registerWorker();
 		}
 
 		// Now try to get them
-		$this->assertEquals($num, count(Resque_Worker::all()));
+		$this->assertEquals($num, count(Worker::all()));
 	}
 
 	public function testGetWorkerById()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
-		$newWorker = Resque_Worker::find((string)$worker);
+		$newWorker = Worker::find((string)$worker);
 		$this->assertEquals((string)$worker, (string)$newWorker);
 	}
 
 	public function testInvalidWorkerDoesNotExist()
 	{
-		$this->assertFalse(Resque_Worker::exists('blah'));
+		$this->assertFalse(Worker::exists('blah'));
 	}
 
 	public function testWorkerCanUnregister()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 		$worker->unregisterWorker();
 
-		$this->assertFalse(Resque_Worker::exists((string)$worker));
-		$this->assertEquals(array(), Resque_Worker::all());
+		$this->assertFalse(Worker::exists((string)$worker));
+		$this->assertEquals(array(), Worker::all());
 		$this->assertEquals(array(), $this->redis->smembers('resque:workers'));
 	}
 
 	public function testPausedWorkerDoesNotPickUpJobs()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->pauseProcessing();
 		Resque::enqueue('jobs', 'Test_Job');
 		$worker->work(0);
@@ -72,8 +72,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testResumedWorkerPicksUpJobs()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->pauseProcessing();
 		Resque::enqueue('jobs', 'Test_Job');
 		$worker->work(0);
@@ -85,11 +85,11 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testWorkerCanWorkOverMultipleQueues()
 	{
-		$worker = new Resque_Worker(array(
+		$worker = new Worker(array(
 			'queue1',
 			'queue2'
 		));
-		$worker->setLogger(new Resque_Log());
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 		Resque::enqueue('queue1', 'Test_Job_1');
 		Resque::enqueue('queue2', 'Test_Job_2');
@@ -103,12 +103,12 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testWorkerWorksQueuesInSpecifiedOrder()
 	{
-		$worker = new Resque_Worker(array(
+		$worker = new Worker(array(
 			'high',
 			'medium',
 			'low'
 		));
-		$worker->setLogger(new Resque_Log());
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
 		// Queue the jobs in a different order
@@ -129,8 +129,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testWildcardQueueWorkerWorksAllQueues()
 	{
-		$worker = new Resque_Worker('*');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('*');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
 		Resque::enqueue('queue1', 'Test_Job_1');
@@ -145,8 +145,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testWorkerDoesNotWorkOnUnknownQueues()
 	{
-		$worker = new Resque_Worker('queue1');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('queue1');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 		Resque::enqueue('queue2', 'Test_Job');
 
@@ -156,8 +156,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 	public function testWorkerClearsItsStatusWhenNotWorking()
 	{
 		Resque::enqueue('jobs', 'Test_Job');
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$job = $worker->reserve();
 		$worker->workingOn($job);
 		$worker->doneWorking();
@@ -166,8 +166,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
 	public function testWorkerRecordsWhatItIsWorkingOn()
 	{
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
 		$payload = array(
@@ -189,8 +189,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 		Resque::enqueue('jobs', 'Test_Job');
 		Resque::enqueue('jobs', 'Invalid_Job');
 
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->work(0);
 		$worker->work(0);
 
@@ -201,59 +201,59 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 	public function testWorkerCleansUpDeadWorkersOnStartup()
 	{
 		// Register a good worker
-		$goodWorker = new Resque_Worker('jobs');
-		$goodWorker->setLogger(new Resque_Log());
+		$goodWorker = new Worker('jobs');
+		$goodWorker->setLogger(new ConsoleLogger());
 		$goodWorker->registerWorker();
 		$workerId = explode(':', $goodWorker);
 
 		// Register some bad workers
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->setId($workerId[0].':1:jobs');
 		$worker->registerWorker();
 
-		$worker = new Resque_Worker(array('high', 'low'));
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker(array('high', 'low'));
+		$worker->setLogger(new ConsoleLogger());
 		$worker->setId($workerId[0].':2:high,low');
 		$worker->registerWorker();
 
-		$this->assertEquals(3, count(Resque_Worker::all()));
+		$this->assertEquals(3, count(Worker::all()));
 
 		$goodWorker->pruneDeadWorkers();
 
 		// There should only be $goodWorker left now
-		$this->assertEquals(1, count(Resque_Worker::all()));
+		$this->assertEquals(1, count(Worker::all()));
 	}
 
 	public function testDeadWorkerCleanUpDoesNotCleanUnknownWorkers()
 	{
 		// Register a bad worker on this machine
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$workerId = explode(':', $worker);
 		$worker->setId($workerId[0].':1:jobs');
 		$worker->registerWorker();
 
 		// Register some other false workers
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->setId('my.other.host:1:jobs');
 		$worker->registerWorker();
 
-		$this->assertEquals(2, count(Resque_Worker::all()));
+		$this->assertEquals(2, count(Worker::all()));
 
 		$worker->pruneDeadWorkers();
 
 		// my.other.host should be left
-		$workers = Resque_Worker::all();
+		$workers = Worker::all();
 		$this->assertEquals(1, count($workers));
 		$this->assertEquals((string)$worker, (string)$workers[0]);
 	}
 
 	public function testWorkerFailsUncompletedJobsOnExit()
 	{
-		$worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+		$worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
 		$worker->registerWorker();
 
 		$payload = array(
@@ -269,8 +269,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
 
     public function testBlockingListPop()
     {
-        $worker = new Resque_Worker('jobs');
-		$worker->setLogger(new Resque_Log());
+        $worker = new Worker('jobs');
+		$worker->setLogger(new ConsoleLogger());
         $worker->registerWorker();
 
         Resque::enqueue('jobs', 'Test_Job_1');
@@ -295,8 +295,8 @@ class Resque_Tests_WorkerTest extends Resque_Tests_TestCase
     {
         Resque::enqueue('jobs', 'Test_Infinite_Recursion_Job');
 
-        $worker = new Resque_Worker('jobs');
-        $worker->setLogger(new Resque_Log());
+        $worker = new Worker('jobs');
+        $worker->setLogger(new ConsoleLogger());
         $worker->work(0);
 
         $this->assertEquals(1, Resque_Stat::get('failed'));